move command hooks code to its own file
